requirements.txt
hydralit_components==1.0.10
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
plotly==5.15.0
Requests==2.31.0
seaborn==0.12.2
streamlit==1.23.1
